move to proper folders, readme update
