+++ docs/process-services-cloud/components/app-list-cloud.component.md
@@ -2,19 +2,19 @@
 Title: App List Cloud Component
 Added: v3.0.0
 Status: Experimental
-Last reviewed: 2019-01-08
+Last reviewed: 2025-01-03
 ---
 
-# [App List Cloud Component](../../../lib/process-services-cloud/src/lib/app/components/app-list-cloud.component.ts "Defined in app-list-cloud.component.ts")
+# App List Cloud Component
+
+`standalone`, `component`
 
 Shows all deployed cloud application instances.
 
 ## Basic Usage
 
 ```html
-<adf-cloud-app-list 
-    [layoutType]="'GRID'">
-</adf-cloud-app-list>
+<adf-cloud-app-list [layoutType]="'GRID'" />
 ```
 
 ### [Transclusions](../../user-guide/transclusion.md)
@@ -23,11 +23,10 @@ You can show custom content when there are no apps available by supplying an
 `<adf-custom-empty-content>` section:
 
 ```html
-<adf-cloud-app-list
-    [layoutType]="'GRID'">
-        <adf-custom-empty-content>
-            No Apps present
-        </adf-custom-empty-content>
+<adf-cloud-app-list [layoutType]="'GRID'">
+    <adf-custom-empty-content>
+      No Apps present
+    </adf-custom-empty-content>
 </adf-cloud-app-list>
 ```
 
@@ -38,19 +37,25 @@ If you are generating a project for Activiti 7, you must add the list of apps yo
 For example :
 
 ```json
-  "alfresco-deployed-apps" : [{"name": "simple-app"}]
+{
+    "alfresco-deployed-apps": [
+        {
+            "name": "simple-app"
+        }
+    ]
+}
 ```
 
 ## Class members
 
 ### Properties
 
-| Name | Type | Default value | Description |
-| ---- | ---- | ------------- | ----------- |
-| layoutType | `string` | LAYOUT_GRID | (**required**) Defines the layout of the apps. There are two possible values, "GRID" and "LIST". |
+| Name       | Type     | Default value | Description                                                                                      |
+|------------|----------|---------------|--------------------------------------------------------------------------------------------------|
+| layoutType | `string` | LAYOUT_GRID   | (**required**) Defines the layout of the apps. There are two possible values, "GRID" and "LIST". |
 
 ### Events
 
-| Name | Type | Description |
-| ---- | ---- | ----------- |
-| appClick | [`EventEmitter`](https://angular.io/api/core/EventEmitter)`<`[`ApplicationInstanceModel`](../../../lib/process-services-cloud/src/lib/app/models/application-instance.model.ts)`>` | Emitted when an app entry is clicked. |
+| Name     | Type                                     | Description                           |
+|----------|------------------------------------------|---------------------------------------|
+| appClick | `EventEmitter<ApplicationInstanceModel>` | Emitted when an app entry is clicked. |
