+++ lib/process-services-cloud/src/lib/app/components/app-details-cloud/app-details-cloud.component.html
