+++ lib/process-services-cloud/src/lib/app/components/app-details-cloud/app-details-cloud.component.scss
@@ -22,16 +22,46 @@
 }
 
 $tile-themes: (
-    theme-1: (bg: #269abc, color: #168aac),
-    theme-2: (bg: #7da9b0, color: #6d99a0),
-    theme-3: (bg: #7689ab, color: #66799b),
-    theme-4: (bg: #c74e3e, color: #b73e2e),
-    theme-5: (bg: #fab96c, color: #eaa95c),
-    theme-6: (bg: #759d4c, color: #658d3c),
-    theme-7: (bg: #b1b489, color: #a1a479),
-    theme-8: (bg: #a17299, color: #916289),
-    theme-9: (bg: #696c67, color: #595c57),
-    theme-10: (bg: #cabb33, color: #baab23)
+    theme-1: (
+        bg: #269abc,
+        color: #168aac
+    ),
+    theme-2: (
+        bg: #7da9b0,
+        color: #6d99a0
+    ),
+    theme-3: (
+        bg: #7689ab,
+        color: #66799b
+    ),
+    theme-4: (
+        bg: #c74e3e,
+        color: #b73e2e
+    ),
+    theme-5: (
+        bg: #fab96c,
+        color: #eaa95c
+    ),
+    theme-6: (
+        bg: #759d4c,
+        color: #658d3c
+    ),
+    theme-7: (
+        bg: #b1b489,
+        color: #a1a479
+    ),
+    theme-8: (
+        bg: #a17299,
+        color: #916289
+    ),
+    theme-9: (
+        bg: #696c67,
+        color: #595c57
+    ),
+    theme-10: (
+        bg: #cabb33,
+        color: #baab23
+    )
 );
 
 adf-cloud-app-details {
@@ -69,10 +99,7 @@ adf-cloud-app-details {
                 max-width: unset;
 
                 &:hover {
-                    box-shadow:
-                        0 8px 10px 1px rgba(0, 0, 0, 0.14),
-                        0 3px 14px 2px rgba(0, 0, 0, 0.12),
-                        0 5px 5px -3px rgba(0, 0, 0, 0.2);
+                    box-shadow: 0 8px 10px 1px rgba(0, 0, 0, 0.14), 0 3px 14px 2px rgba(0, 0, 0, 0.12), 0 5px 5px -3px rgba(0, 0, 0, 0.2);
                     cursor: pointer;
                     transform: scale(1.015);
                 }
