+++ lib/process-services-cloud/src/lib/app/components/app-details-cloud/app-details-cloud.component.spec.ts
@@ -16,10 +16,10 @@
  */
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { fakeApplicationInstance } from '../mock/app-model.mock';
+import { fakeApplicationInstance } from '../../mock/app-model.mock';
 import { AppDetailsCloudComponent } from './app-details-cloud.component';
-import { ProcessServiceCloudTestingModule } from '../../testing/process-service-cloud.testing.module';
-import { DEFAULT_APP_INSTANCE_THEME } from '../models/application-instance.model';
+import { ProcessServiceCloudTestingModule } from '../../../testing/process-service-cloud.testing.module';
+import { DEFAULT_APP_INSTANCE_THEME } from '../../models/application-instance.model';
 
 describe('AppDetailsCloudComponent', () => {
     let component: AppDetailsCloudComponent;
