+++ lib/process-services-cloud/src/lib/app/components/app-list-cloud/app-list-cloud.component.scss
