+++ lib/process-services-cloud/src/lib/app/components/app-list-cloud/app-list-cloud.component.spec.ts
@@ -19,10 +19,10 @@ import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { AlfrescoApiService } from '@alfresco/adf-content-services';
 import { of, throwError } from 'rxjs';
-import { fakeApplicationInstance } from '../mock/app-model.mock';
+import { fakeApplicationInstance } from '../../mock/app-model.mock';
 import { AppListCloudComponent, LAYOUT_GRID, LAYOUT_LIST } from './app-list-cloud.component';
-import { AppsProcessCloudService } from '../services/apps-process-cloud.service';
-import { ProcessServiceCloudTestingModule } from '../../testing/process-service-cloud.testing.module';
+import { AppsProcessCloudService } from '../../services/apps-process-cloud.service';
+import { ProcessServiceCloudTestingModule } from '../../../testing/process-service-cloud.testing.module';
 import { MatIconModule } from '@angular/material/icon';
 import { CustomEmptyContentTemplateDirective } from '@alfresco/adf-core';
